Refactored MNIST wrtier to be namespace only.
