+++ test/io/main.cc
@@ -1,7 +1,7 @@
 #include <fstream>
 #include <iostream>
 
-#include "../../src/io/mnist_reader.h"
+#include "../../src/io/mnist.h"
 
 using namespace neat;
 
@@ -9,14 +9,13 @@ int main(int, char**) {
     // Data set found here: http://yann.lecun.com/exdb/mnist/
 
     MNIST::ImageHeader image_header;
-    MNIST reader;
     std::vector<MNIST::Image> images;
     std::vector<uint> labels;
     const uint n_images(5);
     const std::string file_name_images("/home/lbasov/AI/train-images-idx3-ubyte");
     const std::string file_name_labels("/home/lbasov/AI/train-labels-idx1-ubyte");
 
-    image_header = reader.ReadImageHeader(file_name_images);
+    image_header = MNIST::ReadImageHeader(file_name_images);
 
     std::cout << "N IMAGES READ: " << image_header.n_images << " EXPECTED: 60000" << std::endl;
     std::cout << "N ROWS:        " << image_header.n_rows << " EXPECTED: 28" << std::endl;
@@ -25,7 +24,7 @@ int main(int, char**) {
     std::cout << "------------------------------------------------------------------------" << std::endl;
     std::cout << "READING " + std::to_string(n_images) + " IMAGES" << std::endl;
 
-    images = reader.ReadImages(file_name_images, n_images);
+    images = MNIST::ReadImages(file_name_images, n_images);
 
     for (uint i = 0; i < images.front().pixels.size(); i++) {
         if (images.front().pixels.at(i) > 255) {
@@ -53,7 +52,7 @@ int main(int, char**) {
     std::cout << "------------------------------------------------------------------------" << std::endl;
     std::cout << "READING " + std::to_string(n_images) + " LABELS" << std::endl;
 
-    labels = reader.ReadLabels(file_name_labels, n_images);
+    labels = MNIST::ReadLabels(file_name_labels, n_images);
 
     for (uint p = 0; p < n_images; p++) {
         std::cout << labels.at(p) << std::endl;
